requirements.txt
Flask==3.1.3
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
psycopg[binary]
